fixed self.cards
